Hmem analyzer (#505)
* Experiment runner and analyzer
