+++ examples/kubernetes/workloads/memcached-mutilate/big-wss/kustomization.yaml
@@ -1,5 +1,5 @@
 nameSuffix: -big-wss
-bases:
+resources:
 - ../base
 
 commonLabels:
