+++ examples/kubernetes/workloads/pmbench/kustomization.yaml
@@ -1,6 +1,5 @@
 bases:
 - big
-- big-manually-crirm-annotated
 
 commonLabels:
   fit: 1lm
