+++ examples/kubernetes/workloads/redis-memtier/_hmem/dram-pmem/kustomization.yaml
@@ -1,3 +1,6 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
 nameSuffix: -dram-pmem
 
 resources:
